requirements.txt
matplotlib~=3.6.3
pandas~=1.5.3
torch~=2.0.0
torchvision~=0.15.1
Pillow~=9.4.0
lightning~=2.0.1
numpy~=1.22.0
pytest~=7.1.2
src~=0.0.0
setuptools~=60.2.0